pyproject.toml
[project]
name = "fastapi-init"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi>=0.128.0",
    "fastapi-cache2>=0.2.2",
    "httpx>=0.28.1",
    "prometheus-client>=0.23.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-jose>=3.5.0",
    "redis>=7.1.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.45",
    "structlog>=25.5.0",
    "tenacity>=9.1.2",
    "uvicorn>=0.40.0",
]
